# Make `ParseExternalKey` from the `chunk` package allocations free.