```
~/go/src/github.com/grafana/loki main* ⇣
❯ benchcmp  before.txt after.txt2
benchmark                         old ns/op     new ns/op     delta
Benchmark_ParseExternalKey-16     569           415           -27.03%

benchmark                         old allocs     new allocs     delta
Benchmark_ParseExternalKey-16     2              0              -100.00%

benchmark                         old bytes     new bytes     delta
Benchmark_ParseExternalKey-16     96            0             -100.00%
```

This is a very hot function in Loki being used for every chunk reference we get from the index.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>